refactor(architecture)!: enforce sonar leaf budgets
