+++ integration-tests/src/it/starter/src/main/java/consumer/StarterConsumer.java
@@ -1,9 +1,9 @@
 package consumer;
 
 import io.github.ggomarighetti.jparsqlsearch.compile.SearchCompiler;
-import io.github.ggomarighetti.jparsqlsearch.definition.SearchDefinitionFactory;
+import io.github.ggomarighetti.jparsqlsearch.definition.SearchDefinition;
 
 final class StarterConsumer {
     private SearchCompiler compiler;
-    private SearchDefinitionFactory definitions;
+    private SearchDefinition.Factory definitions;
 }
