+++ jpa-rsql-search-api/src/main/java/io/github/ggomarighetti/jparsqlsearch/exception/SearchDefinitionValidationException.java
@@ -1,5 +1,7 @@
 package io.github.ggomarighetti.jparsqlsearch.exception;
 
+import java.util.Objects;
+
 /** Reports an invalid search definition or incompatible engine/backend setup. */
 public final class SearchDefinitionValidationException extends RuntimeException {
     private static final long serialVersionUID = 1L;
@@ -30,7 +32,7 @@ public final class SearchDefinitionValidationException extends RuntimeException 
      */
     public SearchDefinitionValidationException(String code, String message) {
         super(message);
-        this.code = ValidationExceptionSupport.requireCode(code);
+        this.code = requireCode(code);
     }
 
     /**
@@ -42,7 +44,7 @@ public final class SearchDefinitionValidationException extends RuntimeException 
      */
     public SearchDefinitionValidationException(String code, String message, Throwable cause) {
         super(message, cause);
-        this.code = ValidationExceptionSupport.requireCode(code);
+        this.code = requireCode(code);
     }
 
     /**
@@ -51,6 +53,13 @@ public final class SearchDefinitionValidationException extends RuntimeException 
      * @return stable machine-readable error code
      */
     public String code() {
+        return code;
+    }
+
+    private static String requireCode(String code) {
+        if (Objects.requireNonNull(code, "code must not be null").isBlank()) {
+            throw new IllegalArgumentException("code must not be blank");
+        }
         return code;
     }
 }
