--- jpa-rsql-search-core/src/main/java/io/github/ggomarighetti/jparsqlsearch/compile/SearchCompilationMode.java
@@ -1,6 +1,0 @@
-package io.github.ggomarighetti.jparsqlsearch.compile;
-
-enum SearchCompilationMode {
-    PAGE,
-    SLICE
-}
