+++ jpa-rsql-search-spring-boot-starter/src/test/java/io/github/ggomarighetti/jparsqlsearch/autoconfigure/JpaRsqlSearchRsqlAutoConfigurationTest.java
@@ -13,7 +13,6 @@ import io.github.ggomarighetti.jparsqlsearch.rsql.metadata.RsqlOperatorDescripto
 import io.github.ggomarighetti.jparsqlsearch.rsql.RsqlCompilationRequest;
 import io.github.ggomarighetti.jparsqlsearch.rsql.engine.SearchRsqlEngine;
 import io.github.ggomarighetti.jparsqlsearch.rsql.engine.SearchRsqlEngineBuilder;
-import io.github.ggomarighetti.jparsqlsearch.rsql.engine.SearchRsqlEngineCustomizer;
 import io.github.ggomarighetti.jparsqlsearch.definition.validation.SearchDefinitionValidator;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicReference;
